Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
python-multipart==0.0.9
python-dotenv==1.0.1
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
python-jose==3.3.0
# Use passlib without bcrypt extra; backend hashes with pbkdf2_sha256 only to avoid broken bcrypt wheels.
passlib==1.7.4
slowapi==0.1.9
redis==5.0.8
email-validator==2.2.0
Pillow==10.4.0
# numpy 1.26.x is required for Python 3.12 (requested 1.21.6 is incompatible with >=3.11)
numpy==1.26.4
onnx==1.16.1
onnxruntime-gpu==1.19.2
opencv-python-headless==4.10.0.84
# insightface==0.7.3  # OPTIONAL: heavy dependency; remove comment only if you have build tools or use Python 3.10/3.11 with prebuilt wheels
prometheus-client==0.21.0
pytest==8.3.2
httpx==0.27.2
mediapipe==0.10.14
requests==2.32.3
